fix(types): use signal-channel.ts canonical types, remove inline duplicates
