+++ libs/types/src/index.ts
@@ -65,10 +65,11 @@ export type {
   StatusTransition,
   ExecutionRecord,
   RemediationHistoryEntry,
-  SignalChannel,
-  SignalMetadata,
 } from './feature.js';
 export { normalizeFeatureStatus } from './feature.js';
+
+// Signal channel types (provenance tracking)
+export type { SignalChannel, SignalMetadata } from './signal-channel.js';
 
 // Quarantine types
 export type {
